Upgrade to 0.27.1.

+++ src/dfrouter/RODFDetector.h
@@ -4,7 +4,7 @@
 /// @author  Eric Nicolay
 /// @author  Michael Behrisch
 /// @date    Thu, 16.03.2006
-/// @version $Id: RODFDetector.h 20433 2016-04-13 08:00:14Z behrisch $
+/// @version $Id: RODFDetector.h 21182 2016-07-18 06:46:01Z behrisch $
 ///
 // Class representing a detector within the DFROUTER
 /****************************************************************************/
@@ -168,7 +168,7 @@ public:
     /// @{
 
     bool writeEmitterDefinition(const std::string& file,
-                                const std::map<size_t, RandomDistributor<size_t>* >& dists,
+                                const std::map<int, RandomDistributor<int>* >& dists,
                                 const RODFDetectorFlows& flows,
                                 SUMOTime startTime, SUMOTime endTime, SUMOTime stepOffset,
                                 bool includeUnusedRoutes, SUMOReal scale,
@@ -185,7 +185,7 @@ public:
     void buildDestinationDistribution(const RODFDetectorCon& detectors,
                                       SUMOTime startTime, SUMOTime endTime, SUMOTime stepOffset,
                                       const RODFNet& net,
-                                      std::map<size_t, RandomDistributor<size_t>* >& into) const;
+                                      std::map<int, RandomDistributor<int>* >& into) const;
 
     void computeSplitProbabilities(const RODFNet* net, const RODFDetectorCon& detectors,
                                    const RODFDetectorFlows& flows,
@@ -273,7 +273,7 @@ protected:
     /** @brief Clears the given distributions map, deleting the timed distributions
      * @param[in] dists The distribution map to clear
      */
-    void clearDists(std::map<size_t, RandomDistributor<size_t>* >& dists) const;
+    void clearDists(std::map<int, RandomDistributor<int>* >& dists) const;
 
 
 protected:
