+++ src/duarouter/duarouter_main.cpp
@@ -4,7 +4,7 @@
 /// @author  Jakob Erdmann
 /// @author  Michael Behrisch
 /// @date    Thu, 06 Jun 2002
-/// @version $Id: duarouter_main.cpp 20811 2016-05-30 11:05:58Z behrisch $
+/// @version $Id: duarouter_main.cpp 21202 2016-07-19 13:40:35Z behrisch $
 ///
 // Main for DUAROUTER
 /****************************************************************************/
@@ -106,7 +106,7 @@ computeRoutes(RONet& net, ROLoader& loader, OptionsCont& oc) {
     // prepare the output
     const std::string& filename = oc.getString("output-file");
     std::string altFilename = filename + ".alt";
-    const size_t len = filename.length();
+    const int len = (int)filename.length();
     if (len > 4 && filename.substr(len - 4) == ".xml") {
         altFilename = filename.substr(0, len - 4) + ".alt.xml";
     } else if (len > 4 && filename.substr(len - 4) == ".sbx") {
