+++ src/guisim/GUICalibrator.cpp
@@ -4,7 +4,7 @@
 /// @author  Jakob Erdmann
 /// @author  Michael Behrisch
 /// @date    Mon, 26.04.2004
-/// @version $Id: GUICalibrator.cpp 20433 2016-04-13 08:00:14Z behrisch $
+/// @version $Id: GUICalibrator.cpp 21206 2016-07-20 08:08:35Z behrisch $
 ///
 // Changes flow and speed on a set of lanes (gui version)
 /****************************************************************************/
@@ -362,7 +362,7 @@ GUICalibrator::drawGL(const GUIVisualizationSettings& s) const {
         }
     }
     const SUMOReal exaggeration = s.addSize.getExaggeration(s);
-    for (size_t i = 0; i < myFGPositions.size(); ++i) {
+    for (int i = 0; i < (int)myFGPositions.size(); ++i) {
         const Position& pos = myFGPositions[i];
         SUMOReal rot = myFGRotations[i];
         glPushMatrix();
