+++ src/guisim/GUI_E2_ZS_CollectorOverLanes.h
@@ -4,7 +4,7 @@
 /// @author  Jakob Erdmann
 /// @author  Michael Behrisch
 /// @date    Okt 2003
-/// @version $Id: GUI_E2_ZS_CollectorOverLanes.h 20433 2016-04-13 08:00:14Z behrisch $
+/// @version $Id: GUI_E2_ZS_CollectorOverLanes.h 21182 2016-07-18 06:46:01Z behrisch $
 ///
 // The gui-version of a MS_E2_ZS_CollectorOverLanes.
 /****************************************************************************/
@@ -73,7 +73,7 @@ public:
 
 
 protected:
-    MSE2Collector* buildCollector(size_t c, size_t r,
+    MSE2Collector* buildCollector(int c, int r,
                                   MSLane* l, SUMOReal start, SUMOReal end);
