+++ src/mesosim/MEVehicle.h
@@ -2,7 +2,7 @@
 /// @file    MEVehicle.h
 /// @author  Daniel Krajzewicz
 /// @date    Tue, May 2005
-/// @version $Id: MEVehicle.h 20899 2016-06-07 11:29:08Z namdre $
+/// @version $Id: MEVehicle.h 21182 2016-07-18 06:46:01Z behrisch $
 ///
 // A vehicle from the mesoscopic point of view
 /****************************************************************************/
@@ -201,7 +201,7 @@ public:
      * @param[in] s The current segment
      * @param[in] q The current que
      */
-    inline virtual void setSegment(MESegment* s, size_t idx = 0) {
+    inline virtual void setSegment(MESegment* s, int idx = 0) {
         mySegment = s;
         myQueIndex = idx;
     }
@@ -218,7 +218,7 @@ public:
     /** @brief Returns the index of the que the vehicle is in
      * @return The que index
      */
-    inline size_t getQueIndex() const {
+    inline int getQueIndex() const {
         return myQueIndex;
     }
 
@@ -289,7 +289,7 @@ public:
     }
 
     /// @brief Returns the delay that is accrued due to option --meso-tls-penalty
-    SUMOReal getCurrentTLSPenaltySeconds() const; 
+    SUMOReal getCurrentTLSPenaltySeconds() const;
 
 
     /// Replaces the current route by the given one
@@ -326,7 +326,7 @@ protected:
     MESegment* mySegment;
 
     /// @brief Index of the que the vehicle is in (important for multiqueue extension)
-    size_t myQueIndex;
+    int myQueIndex;
 
     /// @brief The (planned) time of leaving the segment (cell)
     SUMOTime myEventTime;
