+++ src/microsim/MSLane.h
@@ -8,7 +8,7 @@
 /// @author  Michael Behrisch
 /// @author  Mario Krumnow
 /// @date    Mon, 12 Mar 2001
-/// @version $Id: MSLane.h 20863 2016-06-03 09:27:22Z behrisch $
+/// @version $Id: MSLane.h 21201 2016-07-19 11:57:22Z behrisch $
 ///
 // Representation of a lane in the micro simulation
 /****************************************************************************/
@@ -180,7 +180,7 @@ public:
      * @see SUMOVehicleClass
      */
     MSLane(const std::string& id, SUMOReal maxSpeed, SUMOReal length, MSEdge* const edge,
-           unsigned int numericalID, const PositionVector& shape, SUMOReal width,
+           int numericalID, const PositionVector& shape, SUMOReal width,
            SVCPermissions permissions, int index);
 
 
@@ -411,7 +411,7 @@ public:
     /** @brief Returns this lane's numerical id
      * @return This lane's numerical id
      */
-    inline size_t getNumericalID() const {
+    inline int getNumericalID() const {
         return myNumericalID;
     }
 
@@ -607,8 +607,8 @@ public:
     /** @brief Returns the number of stored lanes
      * @return The number of stored lanes
      */
-    static size_t dictSize() {
-        return myDict.size();
+    static int dictSize() {
+        return (int)myDict.size();
     }
 
 
@@ -633,7 +633,7 @@ public:
         Returns the myLinks.end() instead; Further, the number of edges to
         look forward may be given */
     static MSLinkCont::const_iterator succLinkSec(const SUMOVehicle& veh,
-            unsigned int nRouteSuccs,
+            int nRouteSuccs,
             const MSLane& succLinkSource,
             const std::vector<MSLane*>& conts);
 
@@ -985,13 +985,13 @@ protected:
 
     /// @brief detect whether there is a collision between the two vehicles
     bool detectCollisionBetween(SUMOTime timestep, const std::string& stage, const MSVehicle* collider, const MSVehicle* victim,
-                                std::set<const MSVehicle*, SUMOVehicle::ComparatorIdLess>& toRemove, 
+                                std::set<const MSVehicle*, SUMOVehicle::ComparatorIdLess>& toRemove,
                                 std::set<const MSVehicle*>& toTeleport) const;
 
     /// @brief take action upon collision
     void handleCollisionBetween(SUMOTime timestep, const std::string& stage, const MSVehicle* collider, const MSVehicle* victim,
                                 SUMOReal gap, SUMOReal latGap,
-                                std::set<const MSVehicle*, SUMOVehicle::ComparatorIdLess>& toRemove, 
+                                std::set<const MSVehicle*, SUMOVehicle::ComparatorIdLess>& toRemove,
                                 std::set<const MSVehicle*>& toTeleport) const;
 
     /// @brief compute maximum braking distance on this lane
@@ -1012,7 +1012,7 @@ protected:
     SUMOReal basePos(const MSVehicle& veh) const;
 
     /// Unique numerical ID (set on reading by netload)
-    size_t myNumericalID;
+    int myNumericalID;
 
     /// The shape of the lane
     PositionVector myShape;
